StudentClubsPage SSR로 변경
