+++ types/about.ts
@@ -30,3 +30,10 @@ export interface FutureCareers {
     year: number;
   }[];
 }
+
+export interface Club {
+  name: string;
+  engName: string;
+  description: string;
+  imageURL: string;
+}
