--- types/club.ts
@@ -1,6 +1,0 @@
-export interface Club {
-  name: string;
-  engName: string;
-  description: string;
-  imageURL: string;
-}
